requirements.txt
pandas
apache-airflow-providers-postgres
sqlalchemy
psycopg2-binary
requests